refactor: split service.py on modules
